+++ src/store/common/stats.h
@@ -14,7 +14,7 @@ class Stats {
 
   void Increment(const std::string &key, int amount = 1);
   void IncrementList(const std::string &key, size_t idx, int amount = 1);
-  void Add(const std::string &key, int value);
+  void Add(const std::string &key, int64_t value);
   void AddList(const std::string &key, size_t idx, uint64_t value);
 
   void ExportJSON(std::ostream &os);
@@ -23,9 +23,9 @@ class Stats {
 
  private:
   std::mutex mtx;
-  std::unordered_map<std::string, int> statInts;
-  std::unordered_map<std::string, std::vector<int>> statLists;
-  std::unordered_map<std::string, std::vector<int>> statIncLists;
+  std::unordered_map<std::string, int64_t> statInts;
+  std::unordered_map<std::string, std::vector<int64_t>> statLists;
+  std::unordered_map<std::string, std::vector<int64_t>> statIncLists;
   std::unordered_map<std::string, std::vector<std::vector<uint64_t>>> statLoLs;
 };
 #endif /* _STATS_H_ */
